+++ CrowdSource/src/com/CMPUT301F12T07/crowdsource/ViewTaskActivity.java
@@ -44,7 +44,7 @@ public class ViewTaskActivity extends Activity {
         setContentView(R.layout.activity_view_task);
         
         db = new LocalDB(this);
-        this.currentTask = db.getTask(getIntent().getExtras().getInt("taskObject"));
+        this.currentTask = db.getTask(getIntent().getExtras().getLong("taskObject"));
         db.close();
         
         // Getting the task title field
@@ -230,7 +230,7 @@ public class ViewTaskActivity extends Activity {
     public void onResume() {
     	super.onResume();
 
-    	this.currentTask = db.getTask(getIntent().getExtras().getInt("taskObject"));
+    	this.currentTask = db.getTask(getIntent().getExtras().getLong("taskObject"));
     	
     	taskTitle.setText(currentTask.get_title());
     	startDate.setText(currentTask.get_dateCreate());
